chore: move commands to cli package, avoid more than one file on main package
