+++ cmd/sid/main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	sidcli "github.com/babylonchain/staking-indexer/cmd/sid/cli"
 	"github.com/urfave/cli"
 )
 
@@ -16,7 +17,7 @@ func main() {
 	app := cli.NewApp()
 	app.Name = "sid"
 	app.Usage = "Staking Indexer Daemon (sid)."
-	app.Commands = append(app.Commands, startCommand, initCommand)
+	app.Commands = append(app.Commands, sidcli.StartCommand, sidcli.InitCommand)
 
 	if err := app.Run(os.Args); err != nil {
 		fatal(err)
